:art: #310 TOO_MANY_REQUESTS Errorを生成する関数を共通化
