+++ src/app/api/cat-images/validation-results/route.ts
@@ -1,4 +1,5 @@
 import { issueClientCredentialsAccessToken } from '@/api';
+import { createTooManyRequestsError } from '@/app/api/_utils/http';
 import {
   httpStatusCode,
   upstashRedisRestToken,
@@ -62,16 +63,7 @@ const isAcceptableCatImageResponse = (
 export const POST = async (request: NextRequest): Promise<Response> => {
   const { success } = await rateLimit.limit(request.ip ?? 'anonymous');
   if (!success) {
-    const responseBody = {
-      type: 'TOO_MANY_REQUESTS',
-      title: 'Too many requests.',
-      detail:
-        'Too many requests from this IP. Please try again after some time.',
-    };
-
-    const status = httpStatusCode.tooManyRequests;
-
-    return NextResponse.json(responseBody, { status });
+    return createTooManyRequestsError();
   }
 
   const requestBody = (await request.json()) as CatImageValidationRequestBody;
