+++ src/app/api/lgtm-images/recently-created/route.ts
@@ -2,6 +2,7 @@ import {
   isFetchLgtmImagesResponseBody,
   issueClientCredentialsAccessToken,
 } from '@/api';
+import { createTooManyRequestsError } from '@/app/api/_utils/http';
 import {
   httpStatusCode,
   upstashRedisRestToken,
@@ -34,16 +35,7 @@ const rateLimit = new Ratelimit({
 export const GET = async (request: NextRequest): Promise<Response> => {
   const { success } = await rateLimit.limit(request.ip ?? 'anonymous');
   if (!success) {
-    const responseBody = {
-      type: 'TOO_MANY_REQUESTS',
-      title: 'Too many requests.',
-      detail:
-        'Too many requests from this IP. Please try again after some time.',
-    };
-
-    const status = httpStatusCode.tooManyRequests;
-
-    return NextResponse.json(responseBody, { status });
+    return createTooManyRequestsError();
   }
 
   const accessToken = await issueClientCredentialsAccessToken();
